@@ -1,8 +1,8 @@
 module github.com/unistack-org/micro-codec-toml/v3
 
-go 1.15
+go 1.16
 
 require (
 	github.com/BurntSushi/toml v0.3.1
-	github.com/unistack-org/micro/v3 v3.3.17
+	github.com/unistack-org/micro/v3 v3.3.19
 )
